use userdata directory provided by the framework
There is already a framework function that returns the userdata
directory. This function is not only using a system property but also a
default value if the system property is not set. So there is no need for
a null check anymore.
We should use the function provided by the framework.

This has already been fixed in: e72cdf1160c206b53e6aa3f06157ebe219bdcc16
This fix has been reverted by: 796a241ac0d458b38420fec0afb3eb59402fbb72

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>